For synchronize, fix sudo to execute on the remote end of the connection
In 2.0.0.x become was reversed for synchronize. It was happening on the local
machine instead of the remote machine. This restores the ansible-1.9.x
behaviour of doing become on the remote machine. However, there's aspects of
this that are hacky (no hackier than ansible-1.9 but not using 2.0 features).
The big problem is that it does not understand any become method except sudo.
I'm willing to use a partial fix now because we don't want people to get used
to the reversed semantics in their playbooks.